wip-79059 BusinessProject: seperate generate project Invoice in a new service
